+++ src/ui.rs
@@ -1,111 +1,65 @@
 //! Things to do with the "user interface" of the command line tool.
 
-use std::{
-    fmt, num::NonZeroUsize, ops::ControlFlow, process::Stdio, sync::Arc,
-};
+use std::{num::NonZeroUsize, ops::ControlFlow, process::Stdio, sync::Arc};
 
-use crossterm::style::{
-    Attribute, Color, ContentStyle, SetAttribute, Stylize as _,
-};
 use petgraph::graph::{DiGraph, IndexType};
 use tokio::{
     io::{AsyncBufReadExt as _, AsyncRead, BufReader},
     process::Command,
     sync::{Semaphore, mpsc},
 };
+use tracing::Instrument as _;
 
 use crate::{
-    config::{Config, Task},
-    error, graph,
-    name::Named,
+    config::Task, error, graph, name::Named, observability::prelude as o,
+    util::DropGuard,
 };
 
-mod unicode {
-    #![allow(missing_docs)]
-    #![allow(clippy::missing_docs_in_private_items)]
-
-    //! Unicode characters used in the UI.
-
-    pub(crate) const BLACK_LEFT_POINTING: char = '\u{25C0}';
-    pub(crate) const BLACK_RIGHT_POINTING: char = '\u{25B6}';
-    pub(crate) const LIGHT_ARC_DOWN_AND_RIGHT: char = '\u{256D}';
-    pub(crate) const LIGHT_ARC_UP_AND_RIGHT: char = '\u{2570}';
-    pub(crate) const LIGHT_DOWN_AND_HORIZONTAL: char = '\u{252C}';
-    pub(crate) const LIGHT_HORIZONTAL: char = '\u{2500}';
-    pub(crate) const LIGHT_UP_AND_HORIZONTAL: char = '\u{2534}';
-    pub(crate) const LIGHT_VERTICAL: char = '\u{2502}';
-}
-
-/// Distinguish between `stdout` and `stderr`.
-enum StdKind {
+/// Output kind.
+pub(crate) enum OutputKind {
     /// `stdout`.
-    Out,
+    Stdout,
 
     /// `stderr`.
-    Err,
+    Stderr,
 }
 
-/// The sequence of characters to print.
-#[derive(Clone, Copy)]
-pub(crate) enum Sequence {
-    /// The start sequence.
-    Start,
-
-    /// The end sequence.
-    End,
-}
-
-/// Displays a [`Sequence`] given the length of the longest name.
-struct DisplaySequence(Sequence, usize);
-
-impl fmt::Display for DisplaySequence {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let (arc, t, arrow) = match self.0 {
-            Sequence::Start => (
-                unicode::LIGHT_ARC_DOWN_AND_RIGHT,
-                unicode::LIGHT_DOWN_AND_HORIZONTAL,
-                unicode::BLACK_LEFT_POINTING,
-            ),
-            Sequence::End => (
-                unicode::LIGHT_ARC_UP_AND_RIGHT,
-                unicode::LIGHT_UP_AND_HORIZONTAL,
-                unicode::BLACK_RIGHT_POINTING,
-            ),
-        };
-
-        for _ in 0..self.1 {
-            write!(f, " ")?;
-        }
-
-        write!(f, " {arc}{}{t}{arrow}", unicode::LIGHT_HORIZONTAL)
-    }
-}
-
-/// Returns the length of the longest task name.
-#[must_use]
-fn longest_name<I, S>(names: I) -> usize
-where
-    I: IntoIterator<Item = S>,
-    S: AsRef<str>,
-{
-    let mut longest = 0;
-    for name in names {
-        let length = name.as_ref().len();
-
-        if length > longest {
-            longest = length;
+impl OutputKind {
+    /// Convert a string into [`Self`].
+    ///
+    /// # Panics
+    ///
+    /// Panics if the input string is invalid.
+    pub(crate) fn from_str(s: &str) -> Self {
+        match s {
+            "stdout" => Self::Stdout,
+            "stderr" => Self::Stderr,
+            _ => panic!("invalid input"),
         }
     }
 
-    longest
+    /// Get a unique character for the current value.
+    pub(crate) fn to_char(&self) -> char {
+        match self {
+            OutputKind::Stdout => 'O',
+            OutputKind::Stderr => 'E',
+        }
+    }
+}
+
+impl AsRef<str> for OutputKind {
+    fn as_ref(&self) -> &str {
+        match self {
+            Self::Stdout => "stdout",
+            Self::Stderr => "stderr",
+        }
+    }
 }
 
 /// Repeats the output from the `reader` prefixed with the task info.
 async fn repeat_prefixed<R>(
-    longest_prefix: usize,
-    kind: StdKind,
+    kind: OutputKind,
     reader: R,
-    task: Arc<Named<Task>>,
 ) -> Result<(), error::Task>
 where
     R: AsyncRead + Unpin,
@@ -119,17 +73,7 @@ where
         let line = lines.next_line().await.map_err(|e| E::Read(e.into()))?;
 
         if let Some(line) = line {
-            let kind = match kind {
-                StdKind::Out => 'O',
-                StdKind::Err => 'E',
-            };
-
-            println!(
-                "{}{:>longest_prefix$} {sep}{kind}{sep} {line}",
-                SetAttribute(Attribute::Reset),
-                &task.name,
-                sep = unicode::LIGHT_VERTICAL.blue(),
-            );
+            o::info!(kind = AsRef::<str>::as_ref(&kind), data = line, "output");
         } else {
             break;
         }
@@ -142,11 +86,21 @@ where
 /// # Errors
 ///
 /// This can fail for a number of reasons, see [`error::Task`] for details.
-async fn run_task(
-    longest_prefix: usize,
-    task: Arc<Named<Task>>,
-) -> Result<(), error::Task> {
+#[o::instrument(
+    skip_all,
+    fields(
+        // NOTE: Can't just call it `name` because that conflicts with the span
+        // name in tracing-subscriber's JSON output format.
+        task.name = AsRef::<str>::as_ref(&task.name),
+        otel.status_code = o::Empty,
+    ),
+)]
+async fn run_task(task: Arc<Named<Task>>) -> Result<(), error::Task> {
     use error::Task as E;
+
+    let mut otel_status_code = DropGuard::new(o::OtelStatusCode::Error, |x| {
+        x.record(&o::Span::current());
+    });
 
     let command = task
         .value
@@ -163,36 +117,39 @@ async fn run_task(
         .spawn()
         .map_err(|e| E::Spawn(e.into(), command.clone()))?;
 
-    let stdout = tokio::spawn(repeat_prefixed(
-        longest_prefix,
-        StdKind::Out,
-        child.stdout.take().expect("should be able to take child stdout"),
-        task.clone(),
-    ));
+    let stdout = tokio::spawn(
+        repeat_prefixed(
+            OutputKind::Stdout,
+            child.stdout.take().expect("should be able to take child stdout"),
+        )
+        .in_current_span(),
+    );
 
-    let stderr = tokio::spawn(repeat_prefixed(
-        longest_prefix,
-        StdKind::Err,
-        child.stderr.take().expect("should be able to take child stderr"),
-        task.clone(),
-    ));
+    let stderr = tokio::spawn(
+        repeat_prefixed(
+            OutputKind::Stderr,
+            child.stderr.take().expect("should be able to take child stderr"),
+        )
+        .in_current_span(),
+    );
 
     stdout.await.expect("should be able to join stdout")?;
     stderr.await.expect("should be able to join stderr")?;
 
     let status = child.wait().await.map_err(|e| E::Wait(e.into()))?;
 
-    if !status.success() {
-        return Err(E::ExitStatus(status));
+    if status.success() {
+        *otel_status_code = o::OtelStatusCode::Ok;
+        Ok(())
+    } else {
+        Err(E::ExitStatus(status))
     }
-
-    Ok(())
 }
 
 /// Run all tasks in the given graph based on the Engage file.
+#[o::instrument(skip_all, fields(otel.status_code = o::Empty))]
 pub(crate) async fn run_graph<E, Ix>(
     graph: Arc<DiGraph<Arc<Named<Task>>, E, Ix>>,
-    config: Config,
     max_parallelism: Option<NonZeroUsize>,
 ) -> Result<(), error::RunGraph>
 where
@@ -201,17 +158,13 @@ where
 {
     use error::RunGraph as E;
 
-    graph::ensure_acyclic(&graph).map_err(E::Cyclic)?;
-    let longest_name = longest_name(config.tasks.keys());
-    let semaphore = max_parallelism.map(|x| Arc::new(Semaphore::new(x.get())));
+    let mut otel_status_code = DropGuard::new(o::OtelStatusCode::Error, |x| {
+        x.record(&o::Span::current());
+    });
 
-    println!(
-        "{} {}",
-        ContentStyle::new()
-            .with(Color::Blue)
-            .apply(DisplaySequence(Sequence::Start, longest_name)),
-        "starting".blue().bold(),
-    );
+    let span = o::Span::current();
+
+    let semaphore = max_parallelism.map(|x| Arc::new(Semaphore::new(x.get())));
 
     let (error_tx, mut error_rx) = mpsc::channel(16);
     let error_collector = tokio::spawn(async move {
@@ -226,7 +179,9 @@ where
 
     graph::edge_order_par_visit(&graph, {
         let graph = graph.clone();
+        let span = span.clone();
         move |ix| {
+            let _enter = span.enter();
             let task = graph[ix].clone();
             let semaphore = semaphore.clone();
             let error_tx = error_tx.clone();
@@ -242,7 +197,7 @@ where
                     None
                 };
 
-                if let Err(e) = run_task(longest_name, task.clone()).await {
+                if let Err(e) = run_task(task.clone()).await {
                     error_tx
                         .send((task, e))
                         .await
@@ -255,6 +210,7 @@ where
 
                 ControlFlow::Continue(())
             }
+            .in_current_span()
         }
     })
     .await;
@@ -270,26 +226,9 @@ where
         .collect::<Vec<_>>();
 
     if errors.is_empty() {
-        println!(
-            "{}{} {}",
-            SetAttribute(Attribute::Reset),
-            ContentStyle::new()
-                .with(Color::Blue)
-                .apply(DisplaySequence(Sequence::End, longest_name)),
-            "success".bold().green(),
-        );
-
+        *otel_status_code = o::OtelStatusCode::Ok;
         Ok(())
     } else {
-        println!(
-            "{}{} {}",
-            SetAttribute(Attribute::Reset),
-            ContentStyle::new()
-                .with(Color::Blue)
-                .apply(DisplaySequence(Sequence::End, longest_name)),
-            "failure".bold().red(),
-        );
-
-        Err(E::Task(errors))
+        Err(E(errors))
     }
 }